Se añadio edición de protocolo (Pasos): Editar campos y eliminar archivos. Falta Eliminar paso
