+++ controllers/Untitled-1.json
@@ -1,9 +1,44 @@
 [
-    {
-        "ID": 1,
-        "Titulo": "Protocolo de prueba",
-        "Descripcion": 'Protocolo de prueba para probar las funciones del sistema de creacion, edicion de protocolos',
-        "ID_Formulario": 48,
-        "Pasos": [ [Paso], [Paso] ]
-      }
+  {
+    "ID": "32",
+    "Titulo_Paso": "skyshcok@gmail.skyshcok@gmail.comcom",
+    "Titulo_PasoID": "input8",
+    "Descripcion": "skyshcok@gmail.comskyshcok@gmail.com",
+    "DescripcionID": "input9",
+    "DiasEstimados": "1",
+    "DiasEstimadosID": "input10",
+    "AceptaArchivo": 0,
+    "AceptaArchivoName": "input11"
+  },
+  {
+    "ID": "33",
+    "Titulo_Paso": "skyshcok@gmail.comskyshcok@gmail.com",
+    "Titulo_PasoID": "input12",
+    "Descripcion": "skyshcok@gmail.comskyshcok@gmail.com",
+    "DescripcionID": "input13",
+    "DiasEstimados": "1",
+    "DiasEstimadosID": "input14",
+    "AceptaArchivo": 0,
+    "AceptaArchivoName": "input15"
+  },
+  {
+    "Titulo_Paso": "skyshcok@gmail.comskyshcok@gmail.com21",
+    "Titulo_PasoID": "input16",
+    "Descripcion": "skyshcok@gmail.comskyshcok@gmail.com12",
+    "DescripcionID": "input17",
+    "DiasEstimados": "1",
+    "DiasEstimadosID": "input18",
+    "AceptaArchivo": 0,
+    "AceptaArchivoName": "input19"
+  },
+  {
+    "Titulo_Paso": "skyshcok@gmail.com24",
+    "Titulo_PasoID": "input20",
+    "Descripcion": "skyshcok@gmail.com55",
+    "DescripcionID": "input21",
+    "DiasEstimados": "1",
+    "DiasEstimadosID": "input22",
+    "AceptaArchivo": 0,
+    "AceptaArchivoName": "input23"
+  }
 ]
