+++ routes/protocolos.js
@@ -4,6 +4,8 @@ const router = express.Router();
 const ProtocoloController = require("../controllers/ProtocoloController");
 
 router.post("/guardar", ProtocoloController.ProtocoloCrear);
+router.post("/guardarEditar", ProtocoloController.ProtocoloEditarPost);
+
 router.get("/editar/:idProtocolo", ProtocoloController.ProtocoloEditarGet);
 router.get("/", ProtocoloController.list);
 router.get("/reviso", ProtocoloController.reviso);
