+++ views/Protocolo/EditarProtocolo.ejs
@@ -54,13 +54,19 @@
     </style>
     <form class="ProtocoloHeader">
       <div>
-        <input type="text" placeholder="Titulo Protocolo" id="Tprotocolo" />
+        <input
+          type="text"
+          placeholder="Titulo Protocolo"
+          id="Tprotocolo"
+          value="<%-protocolo.Titulo%>"
+        />
       </div>
       <div>
         <input
           type="text"
           placeholder="Descripcion del Protocolo"
           id="Dprotocolo"
+          value="<%-protocolo.Descripcion%>"
         />
       </div>
